Fixed too many opened files error
I've split the test/integration/migrations_test.rb into several files
inside test/integration/migration directory. This will allow us to
keep more easy to read and find specific migration's features per
storage type.

Also, I've added some close/disconnect flow in some tests as you can see below.
This was necessary to help OS flush resources not used anymore.

Later, we can think in separate unit testing from integration one using
sort of Rake task.
