+++ test/repository_test.rb
@@ -27,7 +27,7 @@ describe Lotus::Repository do
         ArticleRepository.clear
       end
 
-      after do
+      after(:each) do
         UserRepository.adapter.disconnect
         ArticleRepository.adapter.disconnect
       end
@@ -454,6 +454,10 @@ describe Lotus::Repository do
       ArticleRepository.clear
 
       ArticleRepository.create(article1)
+    end
+
+    after do
+      UserRepository.adapter.disconnect
     end
 
     describe '.transaction' do
